move tests to commitoffset_ut.cpp
