+++ ydb/core/persqueue/ut/ut_with_sdk/autoscaling_ut.cpp
@@ -622,98 +622,6 @@ Y_UNIT_TEST_SUITE(TopicAutoscaling) {
         readSession2->Close();
     }
 
-    Y_UNIT_TEST(PartitionSplit_OffsetCommit) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-        TTopicClient client = setup.MakeClient();
-
-        setup.Write("message-1", 0);
-        setup.Write("message-2", 0);
-        setup.Write("message-3", 0);
-        setup.Write("message-4", 0);
-        setup.Write("message-5", 0);
-        setup.Write("message-6", 0);
-
-        {
-            ui64 txId = 1006;
-            SplitPartition(setup, ++txId, 0, "a");
-
-            auto describe = setup.DescribeTopic();
-            UNIT_ASSERT_EQUAL(describe.GetPartitions().size(), 3);
-        }
-
-        setup.Write("message-7", 1);
-        setup.Write("message-8", 1);
-        setup.Write("message-9", 1);
-        setup.Write("message-10", 1);
-
-        {
-            ui64 txId = 1007;
-            SplitPartition(setup, ++txId, 1, "0");
-
-            auto describe = setup.DescribeTopic();
-            UNIT_ASSERT_EQUAL(describe.GetPartitions().size(), 5);
-        }
-
-        setup.Write("message-11", 3);
-        setup.Write("message-12", 3);
-
-        auto assertCommittedOffset = [&](size_t partition, size_t expectedOffset, const std::string& msg = "") {
-            auto description = setup.DescribeConsumer();
-            auto stats = description.GetPartitions().at(partition).GetPartitionConsumerStats();
-            UNIT_ASSERT(stats);
-            UNIT_ASSERT_VALUES_EQUAL_C(expectedOffset, stats->GetCommittedOffset(), "Partition " << partition << ": " << msg);
-        };
-
-        {
-            static constexpr size_t commited = 2;
-            auto status = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 1, commited);
-            UNIT_ASSERT(status.IsSuccess());
-
-            assertCommittedOffset(0, 6, "Must be commited to the partition end because it is the parent");
-            assertCommittedOffset(1, commited);
-            assertCommittedOffset(3, 0);
-        }
-
-        {
-            static constexpr size_t commited = 3;
-            auto status = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 0, commited);
-            UNIT_ASSERT(status.IsSuccess());
-
-            assertCommittedOffset(0, commited);
-            assertCommittedOffset(1, 0, "Must be commited to the partition begin because it is the child");
-            assertCommittedOffset(3, 0);
-        }
-    }
-
-    Y_UNIT_TEST(CommitTopPast) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-
-        TTopicClient client = setup.MakeClient();
-
-        setup.Write("message_1", 0);
-        setup.Write("message_2", 0);
-
-        ui64 txId = 1023;
-        SplitPartition(setup, ++txId, 0, "a");
-
-        auto status = client.CommitOffset(TEST_TOPIC, 0, TEST_CONSUMER, 0).GetValueSync();
-        UNIT_ASSERT_VALUES_EQUAL_C(NYdb::EStatus::SUCCESS, status.GetStatus(), "The consumer has just started reading the inactive partition and he can commit");
-
-        status = client.CommitOffset(TEST_TOPIC, 0, TEST_CONSUMER, 1).GetValueSync();
-        UNIT_ASSERT_VALUES_EQUAL_C(NYdb::EStatus::SUCCESS, status.GetStatus(), "A consumer who has not read to the end can commit messages forward.");
-
-        status = client.CommitOffset(TEST_TOPIC, 0, TEST_CONSUMER, 0).GetValueSync();
-        UNIT_ASSERT_VALUES_EQUAL_C(NYdb::EStatus::SUCCESS, status.GetStatus(), "A consumer who has not read to the end can commit messages back.");
-
-        status = client.CommitOffset(TEST_TOPIC, 0, TEST_CONSUMER, 2).GetValueSync();
-        UNIT_ASSERT_VALUES_EQUAL_C(NYdb::EStatus::SUCCESS, status.GetStatus(), "The consumer can commit at the end of the inactive partition.");
-
-        status = client.CommitOffset(TEST_TOPIC, 0, TEST_CONSUMER, 0).GetValueSync();
-        UNIT_ASSERT_VALUES_EQUAL_C(NYdb::EStatus::SUCCESS, status.GetStatus(), "The consumer can commit an offset for inactive, read-to-the-end partitions.");
-    }
-
     Y_UNIT_TEST(ControlPlane_CreateAlterDescribe) {
         auto autoscalingTestTopic = "autoscalit-topic";
         TTopicSdkTestSetup setup = CreateSetup();
@@ -930,407 +838,6 @@ Y_UNIT_TEST_SUITE(TopicAutoscaling) {
         result = client.AlterTopic(autoscalingTestTopic, alterSettings).GetValueSync();
 
         UNIT_ASSERT_VALUES_EQUAL(result.GetStatus(), NYdb::EStatus::BAD_REQUEST);
-    }
-
-    Y_UNIT_TEST(PartitionSplit_DistributedTxCommit) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-        TTopicClient client = setup.MakeClient();
-
-        setup.Write("message-1", 0, "producer-1", 1);
-        setup.Write("message-2", 0, "producer-1", 2);
-        setup.Write("message-3", 0, "producer-1", 3);
-        setup.Write("message-4", 0, "producer-1", 4);
-        setup.Write("message-5", 0, "producer-1", 5);
-        setup.Write("message-6", 0, "producer-2", 6);
-
-        {
-            ui64 txId = 1006;
-            SplitPartition(setup, ++txId, 0, "a");
-
-            auto describe = setup.DescribeTopic();
-            UNIT_ASSERT_EQUAL(describe.GetPartitions().size(), 3);
-        }
-
-        setup.Write("message-7", 1, "producer-1", 7);
-        setup.Write("message-8", 1, "producer-1", 8);
-        setup.Write("message-9", 1, "producer-1", 9);
-        setup.Write("message-10", 1, "producer-2", 10);
-
-        {
-            ui64 txId = 1007;
-            SplitPartition(setup, ++txId, 1, "0");
-
-            auto describe = setup.DescribeTopic();
-            UNIT_ASSERT_EQUAL(describe.GetPartitions().size(), 5);
-        }
-
-        auto count = 0;
-        const auto expected = 10;
-
-        auto result = setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto& x) {
-            auto& messages = x.GetMessages();
-            for (size_t i = 0u; i < messages.size(); ++i) {
-                ++count;
-                auto& message = messages[i];
-                Cerr << "SESSION EVENT read message: " << count << " from partition: " << message.GetPartitionSession()->GetPartitionId() << Endl << Flush;
-                message.Commit();
-            }
-
-            return true;
-        });
-
-        UNIT_ASSERT(result.Timeout);
-        UNIT_ASSERT_VALUES_EQUAL(count, expected);
-
-        auto description = setup.DescribeConsumer();
-        UNIT_ASSERT(description.GetPartitions().size() == 5);
-
-        auto stats1 = description.GetPartitions().at(1).GetPartitionConsumerStats();
-        UNIT_ASSERT(stats1);
-        UNIT_ASSERT(stats1->GetCommittedOffset() == 4);
-    }
-
-    Y_UNIT_TEST(PartitionSplit_DistributedTxCommit_ChildFirst) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-
-        TTopicClient client = setup.MakeClient();
-
-        setup.Write("message-1", 0, "producer-1", 1);
-        setup.Write("message-2", 0, "producer-1", 2);
-        setup.Write("message-3", 0, "producer-1", 3);
-        setup.Write("message-4", 0, "producer-1", 4);
-        setup.Write("message-5", 0, "producer-1", 5);
-        setup.Write("message-6", 0, "producer-2", 6);
-
-        {
-            ui64 txId = 1006;
-            SplitPartition(setup, ++txId, 0, "a");
-
-            auto describe = setup.DescribeTopic();
-            UNIT_ASSERT_EQUAL(describe.GetPartitions().size(), 3);
-        }
-
-        setup.Write("message-7", 1, "producer-1", 7);
-        setup.Write("message-8", 1, "producer-1", 8);
-        setup.Write("message-9", 1, "producer-1", 9);
-        setup.Write("message-10", 1, "producer-2", 10);
-
-        {
-            ui64 txId = 1007;
-            SplitPartition(setup, ++txId, 1, "0");
-
-            auto describe = setup.DescribeTopic();
-            UNIT_ASSERT_EQUAL(describe.GetPartitions().size(), 5);
-        }
-
-
-        auto count = 0;
-        const auto expected = 10;
-
-        std::vector<NYdb::NTopic::TReadSessionEvent::TDataReceivedEvent::TMessage> partition0Messages;
-
-        auto result = setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto& x) {
-            auto& messages = x.GetMessages();
-            for (size_t i = 0u; i < messages.size(); ++i) {
-                auto& message = messages[i];
-                count++;
-                int partitionId = message.GetPartitionSession()->GetPartitionId();
-                Cerr << "SESSION EVENT read message: " << count << " from partition: " << partitionId << Endl << Flush;
-                if (partitionId == 1) {
-                    // Commit messages from partition 1 immediately
-                    message.Commit();
-                } else if (partitionId == 0) {
-                    // Store messages from partition 0 for later
-                    partition0Messages.push_back(message);
-                }
-            }
-
-            return true;
-        });
-
-        UNIT_ASSERT(result.Timeout);
-        UNIT_ASSERT_VALUES_EQUAL(count, expected);
-
-        Sleep(TDuration::Seconds(5));
-
-        {
-            auto description = setup.DescribeConsumer();
-            auto stats = description.GetPartitions().at(1).GetPartitionConsumerStats();
-            UNIT_ASSERT(stats);
-
-            // Messages in the parent partition hasn't been committed
-            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 0);
-        }
-
-        for (auto& message : partition0Messages) {
-            message.Commit();
-        }
-
-        Sleep(TDuration::Seconds(5));
-
-        {
-            auto description = setup.DescribeConsumer();
-            auto stats = description.GetPartitions().at(1).GetPartitionConsumerStats();
-            UNIT_ASSERT(stats);
-
-            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 4);
-        }
-    }
-
-    Y_UNIT_TEST(PartitionSplit_DistributedTxCommit_CheckSessionResetAfterCommit) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-
-        TTopicClient client = setup.MakeClient();
-
-        auto seqNo = 1;
-
-        setup.Write("message-1", 0, "producer-1", seqNo++);
-        setup.Write("message-2", 0, "producer-1", seqNo++);
-        setup.Write("message-3", 0, "producer-1", seqNo++);
-        setup.Write("message-4", 0, "producer-1", seqNo++);
-        setup.Write("message-5", 0, "producer-1", seqNo++);
-        setup.Write("message-6", 0, "producer-2", seqNo++);
-
-        {
-            ui64 txId = 1006;
-            SplitPartition(setup, ++txId, 0, "a");
-
-            auto describe = setup.DescribeTopic();
-            UNIT_ASSERT_EQUAL(describe.GetPartitions().size(), 3);
-        }
-
-        setup.Write("message-7", 1, "producer-2", seqNo++);
-        setup.Write("message-8", 1, "producer-2", seqNo++);
-
-        std::vector<size_t> counters;
-        counters.resize(seqNo - 1);
-
-        auto result = setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto& x) {
-            auto& messages = x.GetMessages();
-            for (size_t i = 0u; i < messages.size(); ++i) {
-                auto& message = messages[i];
-                message.Commit();
-                Cerr << "SESSION EVENT READ SeqNo: " << message.GetSeqNo() << Endl << Flush;
-                auto count = ++counters[message.GetSeqNo() - 1];
-
-                // check we get this SeqNo two times
-                if (message.GetSeqNo() == 6 && count == 1) {
-                    Sleep(TDuration::MilliSeconds(300));
-                    auto status = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 0, 3);
-                    UNIT_ASSERT(status.IsSuccess());
-                }
-            }
-
-            return true;
-        });
-
-        UNIT_ASSERT_VALUES_EQUAL_C(1, counters[0], TStringBuilder() << "Message must be read 1 times because reset commit to offset 3, but 0 message has been read " << counters[0] << " times") ;
-        UNIT_ASSERT_VALUES_EQUAL_C(1, counters[1], TStringBuilder() << "Message must be read 1 times because reset commit to offset 3, but 1 message has been read " << counters[1] << " times") ;
-        UNIT_ASSERT_VALUES_EQUAL_C(1, counters[2], TStringBuilder() << "Message must be read 1 times because reset commit to offset 3, but 2 message has been read " << counters[2] << " times") ;
-
-        UNIT_ASSERT_VALUES_EQUAL_C(2, counters[3], TStringBuilder() << "Message 1 must be read two times, but 3 message has been read " << counters[3] << " times") ;
-        UNIT_ASSERT_VALUES_EQUAL_C(2, counters[4], TStringBuilder() << "Message 1 must be read two times, but 4 message has been read " << counters[4] << " times") ;
-        UNIT_ASSERT_VALUES_EQUAL_C(2, counters[5], TStringBuilder() << "Message 1 must be read two times, but 5 message has been read " << counters[5] << " times") ;
-
-        {
-            auto s = result.StartPartitionSessionEvents[0];
-            UNIT_ASSERT_VALUES_EQUAL(0, s.GetPartitionSession()->GetPartitionId());
-            UNIT_ASSERT_VALUES_EQUAL(0, s.GetCommittedOffset());
-            UNIT_ASSERT_VALUES_EQUAL(6, s.GetEndOffset());
-        }
-        {
-            auto s = result.StartPartitionSessionEvents[3];
-            UNIT_ASSERT_VALUES_EQUAL(0, s.GetPartitionSession()->GetPartitionId());
-            UNIT_ASSERT_VALUES_EQUAL(3, s.GetCommittedOffset());
-            UNIT_ASSERT_VALUES_EQUAL(6, s.GetEndOffset());
-        }
-    }
-
-    Y_UNIT_TEST(PartitionSplit_DistributedTxCommit_CheckOffsetCommitForDifferentCases_SplitedTopic) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        TTopicClient client = setup.MakeClient();
-
-        setup.CreateTopicWithAutoscale();
-
-        auto commit = [&](const std::string& sessionId, ui64 offset) {
-            return setup.Commit(TEST_TOPIC, TEST_CONSUMER, 0, offset, sessionId);
-        };
-
-        auto getConsumerState = [&](ui32 partition) {
-            auto description = setup.DescribeConsumer(TEST_TOPIC, TEST_CONSUMER);
-
-            auto stats = description.GetPartitions().at(partition).GetPartitionConsumerStats();
-            UNIT_ASSERT(stats);
-            return stats;
-        };
-
-        setup.Write("message-1", 0, "producer-1", 1);
-        setup.Write("message-2", 0, "producer-1", 2);
-        setup.Write("message-3", 0, "producer-1", 3);
-        setup.Write("message-4", 0, "producer-1", 4);
-        setup.Write("message-5", 0, "producer-1", 5);
-        setup.Write("message-6", 0, "producer-1", 6);
-        setup.Write("message-7", 0, "producer-1", 7);
-        setup.Write("message-8", 0, "producer-2", 8);
-
-        {
-            ui64 txId = 1006;
-            SplitPartition(setup, ++txId, 0, "a");
-
-            auto describe = client.DescribeTopic(TEST_TOPIC).GetValueSync();
-            UNIT_ASSERT_EQUAL(describe.GetTopicDescription().GetPartitions().size(), 3);
-        }
-
-        setup.Write("message-9", 1, "producer-2", 9);
-        setup.Write("message-10", 1, "producer-2", 10);
-
-        auto commitSent = false;
-        TString readSessionId = "";
-
-        setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto& x) {
-            auto& messages = x.GetMessages();
-            for (size_t i = 0u; i < messages.size(); ++i) {
-                auto& message = messages[i];
-                Cerr << "SESSION EVENT READ SeqNo: " << message.GetSeqNo() << Endl << Flush;
-
-                if (commitSent) {
-                    // read session not changed
-                    UNIT_ASSERT_EQUAL(readSessionId, message.GetPartitionSession()->GetReadSessionId());
-                }
-
-                // check we NOT get this SeqNo two times
-                if (message.GetSeqNo() == 6) {
-                    if (!commitSent) {
-                        commitSent = true;
-
-                        readSessionId = message.GetPartitionSession()->GetReadSessionId();
-
-                        {
-                            auto status = commit(message.GetPartitionSession()->GetReadSessionId(), 8);
-                            UNIT_ASSERT(status.IsSuccess());
-
-                            auto stats = getConsumerState(0);
-                            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 8);
-                        }
-
-                        {
-                            // must be ignored, because commit to past
-                            auto status = commit(message.GetPartitionSession()->GetReadSessionId(), 0);
-                            UNIT_ASSERT(status.IsSuccess());
-
-                            auto stats = getConsumerState(0);
-                            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 8);
-                        }
-
-                        {
-                            // must be ignored, because wrong sessionid
-                            auto status = commit("random session", 0);
-                            UNIT_ASSERT(!status.IsSuccess());
-
-                            Sleep(TDuration::MilliSeconds(500));
-
-                            auto stats = getConsumerState(0);
-                            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 8);
-                        }
-                    } else {
-                        UNIT_ASSERT(false);
-                    }
-                } else {
-                    message.Commit();
-                }
-            }
-
-            return true;
-        });
-    }
-
-    Y_UNIT_TEST(PartitionSplit_DistributedTxCommit_CheckOffsetCommitForDifferentCases_NotSplitedTopic) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-        TTopicClient client = setup.MakeClient();
-
-        auto commit = [&](const std::string& sessionId, ui64 offset) {
-            return setup.Commit(TEST_TOPIC, TEST_CONSUMER, 0, offset, sessionId);
-        };
-
-        auto getConsumerState = [&](ui32 partition) {
-            auto description = setup.DescribeConsumer(TEST_TOPIC, TEST_CONSUMER);
-
-            auto stats = description.GetPartitions().at(partition).GetPartitionConsumerStats();
-            UNIT_ASSERT(stats);
-            return stats;
-        };
-
-        setup.Write("message-1", 0, "producer-1", 1);
-        setup.Write("message-2", 0, "producer-1", 2);
-        setup.Write("message-3", 0, "producer-1", 3);
-        setup.Write("message-4", 0, "producer-1", 4);
-        setup.Write("message-5", 0, "producer-1", 5);
-        setup.Write("message-6", 0, "producer-1", 6);
-        setup.Write("message-7", 0, "producer-1", 7);
-        setup.Write("message-8", 0, "producer-2", 8);
-
-        auto commitSent = false;
-        TString readSessionId = "";
-
-        setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto& x) {
-            auto& messages = x.GetMessages();
-            for (size_t i = 0u; i < messages.size(); ++i) {
-                auto& message = messages[i];
-
-                if (commitSent) {
-                    // read session not changed
-                    UNIT_ASSERT_EQUAL(readSessionId, message.GetPartitionSession()->GetReadSessionId());
-                }
-
-                // check we NOT get this SeqNo two times
-                if (message.GetSeqNo() == 6) {
-                    if (!commitSent) {
-                        commitSent = true;
-                        readSessionId = message.GetPartitionSession()->GetReadSessionId();
-
-                        Sleep(TDuration::MilliSeconds(300));
-
-                        {
-                            auto status = commit(message.GetPartitionSession()->GetReadSessionId(), 8);
-                            UNIT_ASSERT(status.IsSuccess());
-
-                            auto stats = getConsumerState(0);
-                            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 8);
-                        }
-
-                        {
-                            // must be ignored, because commit to past
-                            auto status = commit(message.GetPartitionSession()->GetReadSessionId(), 0);
-                            UNIT_ASSERT(status.IsSuccess());
-
-                            auto stats = getConsumerState(0);
-                            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 8);
-                        }
-
-                        {
-                            // must be ignored, because wrong sessionid
-                            auto status = commit("random session", 0);
-                            UNIT_ASSERT(!status.IsSuccess());
-
-                            Sleep(TDuration::MilliSeconds(500));
-
-                            auto stats = getConsumerState(0);
-                            UNIT_ASSERT_VALUES_EQUAL(stats->GetCommittedOffset(), 8);
-                        }
-                    } else {
-                        UNIT_ASSERT(false);
-                    }
-                } else {
-                    message.Commit();
-                }
-            }
-
-            return true;
-        });
     }
 
     Y_UNIT_TEST(PartitionSplit_AutosplitByLoad) {
