+++ ydb/core/persqueue/ut/ut_with_sdk/topic_ut.cpp
@@ -162,321 +162,6 @@ Y_UNIT_TEST_SUITE(WithSDK) {
             UNIT_ASSERT_VALUES_EQUAL(2, c->GetLastReadOffset());
         }
     }
-
-    void PrepareFlatTopic(TTopicSdkTestSetup& setup) {
-        setup.CreateTopic();
-
-        setup.Write("message-1");
-        setup.Write("message-2");
-        setup.Write("message-3");
-    }
-
-    void PrepareAutopartitionedTopic(TTopicSdkTestSetup& setup) {
-        setup.CreateTopicWithAutoscale();
-
-        // Creating partition hierarchy
-        // 0 ──┬──> 1 ──┬──> 3
-        //     │        └──> 4
-        //     └──> 2
-        //
-        // Each partition has 3 messages
-
-        setup.Write("message-0-1", 0);
-        setup.Write("message-0-2", 0);
-        setup.Write("message-0-3", 0);
-
-        {
-            ui64 txId = 1006;
-            SplitPartition(setup, ++txId, 0, "a");
-        }
-
-        setup.Write("message-1-1", 1);
-        setup.Write("message-1-2", 1);
-        setup.Write("message-1-3", 1);
-
-        setup.Write("message-2-1", 2);
-        setup.Write("message-2-2", 2);
-        setup.Write("message-2-3", 2);
-
-        {
-            ui64 txId = 1007;
-            SplitPartition(setup, ++txId, 1, "0");
-        }
-
-        setup.Write("message-3-1", 3);
-        setup.Write("message-3-2", 3);
-        setup.Write("message-3-3", 3);
-
-        setup.Write("message-4-1", 4);
-        setup.Write("message-4-2", 4);
-        setup.Write("message-4-3", 4);
-    }
-
-    Y_UNIT_TEST(CommitWithWrongSessionId) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        PrepareFlatTopic(setup);
-
-        {
-            auto result = setup.Commit(TString(TEST_TOPIC), TEST_CONSUMER, 0, 1, "wrong-read-session-id");
-            UNIT_ASSERT_C(!result.IsSuccess(), "Commit doesn`t work with wrong session id");
-
-            auto desc = setup.DescribeConsumer(TString(TEST_TOPIC), TEST_CONSUMER);
-            UNIT_ASSERT_VALUES_EQUAL(0, desc.GetPartitions().at(0).GetPartitionConsumerStats()->GetCommittedOffset());
-        }
-    }
-
-    Y_UNIT_TEST(CommitToPastWithWrongSessionId) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        PrepareFlatTopic(setup);
-
-        {
-            auto result = setup.Commit(TString(TEST_TOPIC), TEST_CONSUMER, 0, 2);
-            UNIT_ASSERT_C(result.IsSuccess(), "Commited without session id. It is reset mode");
-
-            auto desc = setup.DescribeConsumer(TString(TEST_TOPIC), TEST_CONSUMER);
-            UNIT_ASSERT_VALUES_EQUAL(2, desc.GetPartitions().at(0).GetPartitionConsumerStats()->GetCommittedOffset());
-        }
-
-        {
-            auto result = setup.Commit(TString(TEST_TOPIC), TEST_CONSUMER, 0, 0, "wrong-read-session-id");
-            UNIT_ASSERT_C(!result.IsSuccess(), "Commit doesn`t work with wrong session id");
-
-            auto desc = setup.DescribeConsumer(TString(TEST_TOPIC), TEST_CONSUMER);
-            UNIT_ASSERT_VALUES_EQUAL(2, desc.GetPartitions().at(0).GetPartitionConsumerStats()->GetCommittedOffset());
-        }
-    }
-
-    Y_UNIT_TEST(Commit_ToParentPartitionWithWrongSessionId) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-
-        setup.Write("message-1", 0);
-
-        {
-            ui64 txId = 1006;
-            SplitPartition(setup, ++txId, 0, "a");
-        }
-
-        setup.Write("message-2", 1);
-
-        Cerr << ">>>>> BEGIN 0" << Endl << Flush;
-        {
-            auto result = setup.Commit(TString(TEST_TOPIC), TEST_CONSUMER, 0, 1);
-            UNIT_ASSERT_C(result.IsSuccess(), "Commited without session id. It is reset mode");
-
-            auto desc = setup.DescribeConsumer(TString(TEST_TOPIC), TEST_CONSUMER);
-            UNIT_ASSERT_VALUES_EQUAL(1, desc.GetPartitions().at(0).GetPartitionConsumerStats()->GetCommittedOffset());
-        }
-
-        Cerr << ">>>>> BEGIN 1" << Endl << Flush;
-        {
-            auto result = setup.Commit(TString(TEST_TOPIC), TEST_CONSUMER, 1, 1);
-            UNIT_ASSERT_C(result.IsSuccess(), "Commited without session id. It is reset mode");
-
-            auto desc = setup.DescribeConsumer(TString(TEST_TOPIC), TEST_CONSUMER);
-            UNIT_ASSERT_VALUES_EQUAL(1, desc.GetPartitions().at(1).GetPartitionConsumerStats()->GetCommittedOffset());
-        }
-
-        Cerr << ">>>>> BEGIN 2" << Endl << Flush;
-        {
-            auto result = setup.Commit(TString(TEST_TOPIC), TEST_CONSUMER, 0, 0, "wrong-read-session-id");
-            UNIT_ASSERT_C(!result.IsSuccess(), "Commit doesn`t work with wrong session id");
-
-            auto desc = setup.DescribeConsumer(TString(TEST_TOPIC), TEST_CONSUMER);
-            UNIT_ASSERT_VALUES_EQUAL_C(1, desc.GetPartitions().at(0).GetPartitionConsumerStats()->GetCommittedOffset(), "Offset doesn`t changed");
-        }
-        Cerr << ">>>>> END" << Endl << Flush;
-
-    }
-
-    Y_UNIT_TEST(Commit_WithoutSession_ParentNotFinished) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        PrepareAutopartitionedTopic(setup);
-
-        auto getCommittedOffset = [&](size_t partition) {
-            auto desc = setup.DescribeConsumer();
-            return desc.GetPartitions().at(partition).GetPartitionConsumerStats()->GetCommittedOffset();
-        };
-
-        {
-            // Commit parent partition to non end
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 0, 1);
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-
-        {
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 1, 1);
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-    }
-
-    Y_UNIT_TEST(Commit_WithoutSession_ToPastParentPartition) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        PrepareAutopartitionedTopic(setup);
-
-        auto getCommittedOffset = [&](size_t partition) {
-            auto desc = setup.DescribeConsumer();
-            return desc.GetPartitions().at(partition).GetPartitionConsumerStats()->GetCommittedOffset();
-        };
-
-        {
-            // Commit child partition to non end
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 3, 1);
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-
-        {
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 1, 1);
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-    }
-
-    Y_UNIT_TEST(Commit_WithSession_ParentNotFinished) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        PrepareAutopartitionedTopic(setup);
-
-        auto getCommittedOffset = [&](size_t partition) {
-            auto desc = setup.DescribeConsumer();
-            return desc.GetPartitions().at(partition).GetPartitionConsumerStats()->GetCommittedOffset();
-        };
-
-        {
-            // Commit parent partition to non end
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 0, 1);
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-
-        {
-            auto r = setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto&) {
-                return false;
-            });
-
-            // Commit parent to middle
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 1, 1, r.StartPartitionSessionEvents.front().GetPartitionSession()->GetReadSessionId());
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-    }
-
-    Y_UNIT_TEST(Commit_WithSession_ToPastParentPartition) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        PrepareAutopartitionedTopic(setup);
-
-        auto getCommittedOffset = [&](size_t partition) {
-            auto desc = setup.DescribeConsumer();
-            return desc.GetPartitions().at(partition).GetPartitionConsumerStats()->GetCommittedOffset();
-        };
-
-        {
-            // Commit parent partition to non end
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 3, 1);
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-
-        {
-            auto r = setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto&) {
-                return false;
-            });
-
-            // Commit parent to middle
-            auto result = setup.Commit(TEST_TOPIC, TEST_CONSUMER, 1, 1, r.StartPartitionSessionEvents.front().GetPartitionSession()->GetReadSessionId());
-            UNIT_ASSERT(result.IsSuccess());
-
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(3, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-            UNIT_ASSERT_VALUES_EQUAL(1, getCommittedOffset(3));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(4));
-        }
-    }
-
-    Y_UNIT_TEST(Commit_ToNewChild_WithoutCommitToParent) {
-        TTopicSdkTestSetup setup = CreateSetup();
-        setup.CreateTopicWithAutoscale();
-
-        setup.Write("message-0-0", 0);
-        setup.Write("message-0-1", 0);
-        setup.Write("message-0-2", 0);
-
-        auto getCommittedOffset = [&](size_t partition) {
-            auto desc = setup.DescribeConsumer();
-            return desc.GetPartitions().at(partition).GetPartitionConsumerStats()->GetCommittedOffset();
-        };
-
-        {
-            bool committed = false;
-
-            auto r = setup.Read(TEST_TOPIC, TEST_CONSUMER, [&](auto& x) {
-                for (auto & m: x.GetMessages()) {
-                    if (x.GetPartitionSession()->GetPartitionId() == 0 && m.GetOffset() == 1) {
-                        ui64 txId = 1006;
-                        SplitPartition(setup, ++txId, 0, "a");
-
-                        setup.Write("message-1-0", 1);
-                        setup.Write("message-1-1", 1);
-                        setup.Write("message-1-2", 1);
-                    } else if (x.GetPartitionSession()->GetPartitionId() == 1 && m.GetOffset() == 0) {
-                        m.Commit();
-                        committed = true;
-                        return false;
-                    }
-                }
-
-                return true;
-            });
-
-            UNIT_ASSERT(committed);
-
-            Sleep(TDuration::Seconds(3));
-
-            // Commit hasn`t applyed because messages from the parent partitions has`t been committed
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(0));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(1));
-            UNIT_ASSERT_VALUES_EQUAL(0, getCommittedOffset(2));
-        }
-    }
-
 }
 
 } // namespace NKikimr
