+++ ydb/core/persqueue/ut/ut_with_sdk/ya.make
@@ -31,6 +31,7 @@ YQL_LAST_ABI_VERSION()
 SRCS(
     autoscaling_ut.cpp
     balancing_ut.cpp
+    commitoffset_ut.cpp
     mirrorer_ut.cpp
     topic_ut.cpp
 )
